+++ sdk/src/spectrumx/__init__.py
@@ -37,22 +37,32 @@ __version__ = importlib.metadata.version(LIB_NAME)
 
 
 class Experiments:
-    def enable_capture_search(self) -> None:
-        """Enables the experimental capture search feature.
+    def enable_capture_advanced_search(self) -> None:
+        """Enables the experimental advanced capture search feature.
 
         This feature is experimental and may change or be removed in future versions.
+
+        Usage after activation:
+
+        ```
+        from spectrumx import client
+        sds = client(...)
+        sds.captures.advanced_search(...)
+
+        # to read docstring with usage:
+        print(sds.captures.advanced_search.__doc__)
+        ```
+
         """
         try:
             from spectrumx.api.captures import (
-                _enable_experimental_search,  # pyright: ignore[reportPrivateUsage]
+                _enable_experimental_advanced_search,  # pyright: ignore[reportPrivateUsage]
             )
 
-            _enable_experimental_search()
+            _enable_experimental_advanced_search()
         except ImportError as err:
-            log_user_error(
-                "Failed to import OpenSearch DSL module. Ensure it is installed."
-            )
-            msg = "OpenSearch DSL module is required for this feature."
+            msg = "Failed to import the experimental advanced search."
+            log_user_error(msg)
             raise ImportError(msg) from err